apps/serve/requirements.txt
# Hominem Serving System Requirements

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
# sqlite3 is built-in to Python - no installation needed

# Environment variables
python-dotenv>=1.0.0

# Model loading for LoRA integration
transformers>=4.35.0
torch>=2.1.0
peft>=0.6.0
accelerate>=0.24.0  # For device mapping
